Rewrite to TypeScript
Use lts node
