+++ .gitignore
@@ -44,3 +44,5 @@ jspm_packages
 .DS_Store
 
 .history/
+
+lib
