--- index.js
@@ -1,3 +1,0 @@
-module.exports.FixedThreadPool = require('./lib/fixed')
-module.exports.DynamicThreadPool = require('./lib/dynamic')
-module.exports.ThreadWorker = require('./lib/workers').ThreadWorker
