+++ package.json
@@ -2,11 +2,10 @@
   "name": "poolifier",
   "version": "1.2.1",
   "description": "Library on top of node js worker threads that implement various worker pools type",
-  "main": "index.js",
-  "types": "index.d.ts",
+  "main": "lib/index.js",
   "scripts": {
-    "build": "npm install",
-    "test": "standard && nyc mocha --exit --timeout 20000 tests/*test.js ",
+    "build": "npm install && rimraf lib && tsc",
+    "test": "rimraf lib && tsc && nyc mocha --exit --timeout 20000 tests/*test.js ",
     "debug-test": "mocha --inspect-brk --exit tests/*test.js ",
     "demontest": "nodemon --exec \"npm test\"",
     "coverage": "nyc report --reporter=text-lcov | coveralls",
@@ -46,12 +45,12 @@
     "mocha-lcov-reporter": "^1.3.0",
     "nyc": "^15.0.0",
     "prettier-standard": "^16.4.1",
-    "standard": "^16.0.0"
+    "rimraf": "^3.0.2",
+    "standard": "^16.0.0",
+    "typescript": "^4.1.3"
   },
   "engines": {
     "node": ">=12.0.0"
   },
-  "dependencies": {
-    "typescript": "^4.1.3"
-  }
+  "dependencies": {}
 }
