+++ src/dynamic.ts
@@ -1,0 +1,71 @@
+import { EventEmitter } from 'events'
+import type { FixedThreadPoolOptions, WorkerWithMessageChannel } from './fixed'
+import FixedThreadPool from './fixed'
+
+class MyEmitter extends EventEmitter {}
+
+export type DynamicThreadPoolOptions = FixedThreadPoolOptions
+
+/**
+ * A thread pool with a min/max number of threads, is possible to execute tasks in sync or async mode as you prefer.
+ *
+ * This thread pool will create new workers when the other ones are busy, until the max number of threads,
+ * when the max number of threads is reached, an event will be emitted, if you want to listen this event use the emitter method.
+ *
+ * @author [Alessandro Pio Ardizio](https://github.com/pioardi)
+ * @since 0.0.1
+ */
+export default class DynamicThreadPool<
+  Data = any,
+  Response = any
+> extends FixedThreadPool<Data, Response> {
+  public readonly emitter: MyEmitter
+
+  /**
+   * @param min Min number of threads that will be always active
+   * @param max Max number of threads that will be active
+   * @param filename A file path with implementation of `ThreadWorker` class, relative path is fine.
+   * @param opts An object with possible options for example `errorHandler`, `onlineHandler`. Default: `{ maxTasks: 1000 }`
+   */
+  public constructor (
+    public readonly min: number,
+    public readonly max: number,
+    public readonly filename: string,
+    public readonly opts: DynamicThreadPoolOptions = { maxTasks: 1000 }
+  ) {
+    super(min, filename, opts)
+
+    this.emitter = new MyEmitter()
+  }
+
+  protected _chooseWorker (): WorkerWithMessageChannel {
+    let worker: WorkerWithMessageChannel
+    for (const entry of this.tasks) {
+      if (entry[1] === 0) {
+        worker = entry[0]
+        break
+      }
+    }
+
+    if (worker) {
+      // a worker is free, use it
+      return worker
+    } else {
+      if (this.workers.length === this.max) {
+        this.emitter.emit('FullPool')
+        return super._chooseWorker()
+      }
+      // all workers are busy create a new worker
+      const worker = this._newWorker()
+      worker.port2.on('message', (message) => {
+        if (message.kill) {
+          worker.postMessage({ kill: 1 })
+          worker.terminate()
+        }
+      })
+      return worker
+    }
+  }
+}
+
+module.exports = DynamicThreadPool
