+++ src/workers.ts
@@ -1,18 +1,43 @@
-'use strict'
-const { isMainThread, parentPort } = require('worker_threads')
-const { AsyncResource } = require('async_hooks')
+import { isMainThread, parentPort } from 'worker_threads'
+import { AsyncResource } from 'async_hooks'
+
+export interface ThreadWorkerOptions {
+  /**
+   * Max time to wait tasks to work on (in ms), after this period the new worker threads will die.
+   *
+   * @default 60.000 ms
+   */
+  maxInactiveTime?: number
+  /**
+   * `true` if your function contains async pieces, else `false`.
+   *
+   * @default false
+   */
+  async?: boolean
+}
 
 /**
- * An example worker that will be always alive, you just need to extend this class if you want a static pool.<br>
- * When this worker is inactive for more than 1 minute, it will send this info to the main thread,<br>
- * if you are using DynamicThreadPool, the workers created after will be killed, the min num of thread will be guaranteed
- * @author Alessandro Pio Ardizio
+ * An example worker that will be always alive, you just need to **extend** this class if you want a static pool.
+ *
+ * When this worker is inactive for more than 1 minute, it will send this info to the main thread,
+ * if you are using DynamicThreadPool, the workers created after will be killed, the min num of thread will be guaranteed.
+ *
+ * @author [Alessandro Pio Ardizio](https://github.com/pioardi)
  * @since 0.0.1
  */
-class ThreadWorker extends AsyncResource {
-  constructor (fn, opts) {
+export class ThreadWorker<Data = any, Response = any> extends AsyncResource {
+  protected readonly maxInactiveTime: number
+  protected readonly async: boolean
+  protected lastTask: number
+  protected readonly interval: NodeJS.Timeout
+  protected parent: any
+
+  public constructor (
+    fn: (data: Data) => Response,
+    public readonly opts: ThreadWorkerOptions = {}
+  ) {
     super('worker-thread-pool:pioardi')
-    this.opts = opts || {}
+
     this.maxInactiveTime = this.opts.maxInactiveTime || 1000 * 60
     this.async = !!this.opts.async
     this.lastTask = Date.now()
@@ -25,7 +50,7 @@ class ThreadWorker extends AsyncResource {
       )
       this._checkAlive.bind(this)()
     }
-    parentPort.on('message', value => {
+    parentPort.on('message', (value) => {
       if (value && value.data && value._id) {
         // here you will receive messages
         // console.log('This is the main thread ' + isMainThread)
@@ -46,13 +71,16 @@ class ThreadWorker extends AsyncResource {
     })
   }
 
-  _checkAlive () {
+  protected _checkAlive (): void {
     if (Date.now() - this.lastTask > this.maxInactiveTime) {
       this.parent.postMessage({ kill: 1 })
     }
   }
 
-  _run (fn, value) {
+  protected _run (
+    fn: (data: Data) => Response,
+    value: { readonly data: Data, readonly _id: number }
+  ): void {
     try {
       const res = fn(value.data)
       this.parent.postMessage({ data: res, _id: value._id })
@@ -63,17 +91,18 @@ class ThreadWorker extends AsyncResource {
     }
   }
 
-  _runAsync (fn, value) {
+  protected _runAsync (
+    fn: (data: Data) => Promise<Response>,
+    value: { readonly data: Data, readonly _id: number }
+  ): void {
     fn(value.data)
-      .then(res => {
+      .then((res) => {
         this.parent.postMessage({ data: res, _id: value._id })
         this.lastTask = Date.now()
       })
-      .catch(e => {
+      .catch((e) => {
         this.parent.postMessage({ error: e, _id: value._id })
         this.lastTask = Date.now()
       })
   }
 }
-
-module.exports.ThreadWorker = ThreadWorker
